% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-and-helpers.R
\name{pal_twrite}
\alias{pal_twrite}
\title{Twrite Discrete Color Palettes}
\usage{
pal_twrite(
  palette = c("default", "sex", "region", "cat5", "cat10"),
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{palette}{Palette type.
Currently there are four available general options:
\code{"default", "contrast", "GnRd", "PuBl"} containing all or a subset of the palette
(The default package is a 14-color palette inspired by \emph{Twrite}).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{reverse}{Logical. Should the order of the colors be reversed?}
}
\description{
Color palettes inspired by the housestyle of Twrite
}
\examples{
\dontrun{library("scales")
show_col(pal_twrite("default")(14))
show_col(pal_twrite("default", alpha = 0.6)(14))}
}
