% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-and-helpers.R
\name{pal_twrite_c}
\alias{pal_twrite_c}
\title{Twrite Continuous Color Palettes}
\usage{
pal_twrite_c(
  palette = c("red", "green", "red-yellow-green"),
  n = 10,
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{palette}{Palette type.
There are 3 available options:
\itemize{
\item \code{"red"}
\item \code{"green"}
\item \code{"red-yellow-green"}}}

\item{n}{Number of individual colors to be generated.}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{reverse}{Logical. Should the order of the colors be reversed?}
}
\description{
Twrite Continuous Color Palettes
}
\examples{
\dontrun{library("scales")
show_col(pal_twrite("red-yellow-green")(10))
show_col(pal_twrite("red-yellow-green", n = 30, alpha = 0.6, reverse = TRUE)(30))}
}
