% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-and-helpers.R
\name{twrite_style}
\alias{twrite_style}
\title{Add twrite theme to ggplot chart}
\usage{
twrite_style(font = "Arial")
}
\arguments{
\item{font}{Specify one of the available fonts (Default: Arial)}
}
\description{
This function allows you to add the twrite theme to your ggplot-graphics.
}
\examples{
\dontrun{line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 75, size = 1, colour="#3AAA35") +
twrite_style()}
}
\keyword{twrite_style}
