% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twriteproject.R
\name{twriteproject}
\alias{twriteproject}
\title{Create Twrite project folder and file structure}
\usage{
twriteproject()
}
\value{
Folder and file structure for a T-write project
}
\description{
The \code{twriteproject} function will create the folder and file structure
of a T-write project.
}
\examples{
\dontrun{
twriteproject()
}
}
